Rename `commodore/git.py` to `commodore/gitrepo.py`
