+++ commodore/compile.py
@@ -1,6 +1,6 @@
 import click
 
-from . import git
+from . import gitrepo
 from .catalog import fetch_customer_catalog, clean_catalog, update_catalog
 from .cluster import (
     Cluster,
@@ -29,14 +29,14 @@ from .refs import update_refs
 
 def _fetch_global_config(cfg: Config, cluster: Cluster):
     click.secho("Updating global config...", bold=True)
-    repo = git.clone_repository(
+    repo = gitrepo.clone_repository(
         cluster.global_git_repo_url, cfg.inventory.global_config_dir, cfg
     )
     rev = cluster.global_git_repo_revision
     if cfg.global_repo_revision_override:
         rev = cfg.global_repo_revision_override
     if rev:
-        git.checkout_version(repo, rev)
+        gitrepo.checkout_version(repo, rev)
     cfg.register_config("global", repo)
 
 
@@ -45,14 +45,14 @@ def _fetch_customer_config(cfg: Config, cluster: Cluster):
     repo_url = cluster.config_repo_url
     if cfg.debug:
         click.echo(f" > Cloning customer config {repo_url}")
-    repo = git.clone_repository(
+    repo = gitrepo.clone_repository(
         repo_url, cfg.inventory.tenant_config_dir(cluster.tenant_id), cfg
     )
     rev = cluster.config_git_repo_revision
     if cfg.tenant_repo_revision_override:
         rev = cfg.tenant_repo_revision_override
     if rev:
-        git.checkout_version(repo, rev)
+        gitrepo.checkout_version(repo, rev)
     cfg.register_config("customer", repo)
 
 
@@ -104,10 +104,10 @@ def _local_setup(config: Config, cluster_id):
 
     click.secho("Registering config...", bold=True)
     config.register_config(
-        "global", git.init_repository(config.inventory.global_config_dir)
+        "global", gitrepo.init_repository(config.inventory.global_config_dir)
     )
     config.register_config(
-        "customer", git.init_repository(config.inventory.tenant_config_dir(tenant))
+        "customer", gitrepo.init_repository(config.inventory.tenant_config_dir(tenant))
     )
 
     click.secho("Resetting targets...", bold=True)
@@ -127,7 +127,7 @@ def _local_setup(config: Config, cluster_id):
     update_target(config, config.inventory.bootstrap_target)
 
     click.secho("Configuring catalog repo...", bold=True)
-    return git.init_repository(config.catalog_dir)
+    return gitrepo.init_repository(config.catalog_dir)
 
 
 def check_parameters_component_versions(config: Config, cluster_parameters):
